wip(api): Move api classes to 'draft_api.py'
The only remaining classes in 'api.py' are the ones used by
'iob_and.py'.

Rename 'interfaces.py' to 'iob_interface.py';
Rename 'interface' class to 'iob_interface';
